move expression generation to common
