+++ x-pack/plugins/lens/common/get_datasource_layers.ts
@@ -1,0 +1,34 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import memoizeOne from 'memoize-one';
+import type { DatasourceStates, DataViewsState } from '../public/state_management';
+import type { DatasourceLayers, DatasourceMap } from '../public/types';
+
+export const getDatasourceLayers = memoizeOne(function getDatasourceLayers(
+  datasourceStates: DatasourceStates,
+  datasourceMap: DatasourceMap,
+  indexPatterns: DataViewsState['indexPatterns']
+) {
+  const datasourceLayers: DatasourceLayers = {};
+  Object.keys(datasourceMap)
+    .filter((id) => datasourceStates[id] && !datasourceStates[id].isLoading)
+    .forEach((id) => {
+      const datasourceState = datasourceStates[id].state;
+      const datasource = datasourceMap[id];
+
+      const layers = datasource.getLayers(datasourceState);
+      layers.forEach((layer) => {
+        datasourceLayers[layer] = datasourceMap[id].getPublicAPI({
+          state: datasourceState,
+          layerId: layer,
+          indexPatterns,
+        });
+      });
+    });
+  return datasourceLayers;
+});
