+++ x-pack/plugins/lens/public/app_plugin/shared/edit_on_the_fly/layer_configuration_section.tsx
@@ -10,7 +10,7 @@ import { EuiSpacer, useEuiTheme } from '@elastic/eui';
 import { css } from '@emotion/react';
 import { VisualizationToolbar } from '../../../editor_frame_service/editor_frame/workspace_panel';
 import { ConfigPanelWrapper } from '../../../editor_frame_service/editor_frame/config_panel/config_panel';
-import { createIndexPatternService } from '../../../data_views_service/service';
+import { createIndexPatternService } from '../../../../common/data_views_service/service';
 import { useLensDispatch, updateIndexPatterns } from '../../../state_management';
 import { replaceIndexpattern } from '../../../state_management/lens_slice';
 import type { LayerConfigurationProps } from './types';
