+++ x-pack/plugins/lens/public/datasources/form_based/dimension_panel/droppable/on_drop_handler.test.ts
@@ -10,7 +10,7 @@ import { onDrop } from './on_drop_handler';
 import { FormBasedPrivateState } from '../../types';
 import { OperationMetadata, DatasourceDimensionDropHandlerProps } from '../../../../types';
 import { FormulaIndexPatternColumn, MedianIndexPatternColumn } from '../../operations';
-import { generateId } from '../../../../id_generator';
+import { generateId } from '../../../../../common/id_generator';
 import {
   mockDataViews,
   mockedLayers,
