+++ x-pack/plugins/lens/public/datasources/form_based/operations/definitions/formula/formula_public_api.ts
@@ -8,7 +8,7 @@
 import type { DataView } from '@kbn/data-views-plugin/public';
 import { Query } from '@kbn/es-query';
 import type { DateRange } from '../../../../../../common/types';
-import { convertDataViewIntoLensIndexPattern } from '../../../../../data_views_service/loader';
+import { convertDataViewIntoLensIndexPattern } from '../../../../../../common/data_views_service/loader';
 import type { IndexPattern } from '../../../../../types';
 import type { PersistedIndexPatternLayer } from '../../../types';
 import type { TimeScaleUnit } from '../../../../../../common/expressions';
