+++ x-pack/plugins/lens/public/datasources/text_based/text_based_languages.test.ts
@@ -12,7 +12,7 @@ import { TextBasedPersistedState, TextBasedPrivateState } from './types';
 import { dataPluginMock } from '@kbn/data-plugin/public/mocks';
 import { dataViewPluginMocks } from '@kbn/data-views-plugin/public/mocks';
 import { getTextBasedDatasource } from './text_based_languages';
-import { generateId } from '../../id_generator';
+import { generateId } from '../../../common/id_generator';
 import { DatasourcePublicAPI, Datasource, FramePublicAPI } from '../../types';
 
 jest.mock('../../id_generator');
