+++ x-pack/plugins/lens/public/datasources/text_based/text_based_languages.tsx
@@ -33,7 +33,7 @@ import {
   DataSourceInfo,
   UserMessage,
 } from '../../types';
-import { generateId } from '../../id_generator';
+import { generateId } from '../../../common/id_generator';
 import type {
   TextBasedPrivateState,
   TextBasedPersistedState,
