+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/layer_panel.test.tsx
@@ -13,7 +13,7 @@ import { FramePublicAPI, Visualization, VisualizationConfigProps } from '../../.
 import { LayerPanel } from './layer_panel';
 import { LayerActions } from './layer_actions';
 import { coreMock } from '@kbn/core/public/mocks';
-import { generateId } from '../../../id_generator';
+import { generateId } from '../../../../common/id_generator';
 import {
   createMockVisualization,
   createMockFramePublicAPI,
