+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/workspace_panel/chart_switch.tsx
@@ -42,7 +42,7 @@ import {
   selectVisualization,
   selectDatasourceStates,
 } from '../../../state_management';
-import { generateId } from '../../../id_generator/id_generator';
+import { generateId } from '../../../../common/id_generator/id_generator';
 
 interface VisualizationSelection {
   visualizationId: string;
