+++ x-pack/plugins/lens/public/id_generator.ts
@@ -4,6 +4,5 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-export { DOCUMENT_FIELD_NAME } from './constants';
-export type { PersistableFilter, LegacyMetricState } from './types';
-export { docToExpression } from './doc_to_expression';
+
+export { generateId } from '../common/id_generator';
