+++ x-pack/plugins/lens/public/visualizations/metric/visualization.tsx
@@ -31,7 +31,7 @@ import {
 import { GROUP_ID, LENS_METRIC_ID } from './constants';
 import { DimensionEditor, DimensionEditorAdditionalSection } from './dimension_editor';
 import { Toolbar } from './toolbar';
-import { generateId } from '../../id_generator';
+import { generateId } from '../../../common/id_generator';
 import { toExpression } from './to_expression';
 import { nonNullable } from '../../utils';
